+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# [1.0.15] - 2017-12-21
+## Changed
+- DeviceDao.getDeviceList: add project filter
+
+## Fixed
+- TagDao.getTag
+
 # [1.0.14] - 2017-12-19
 ## Added
 - DeviceDao